+++ Documentation/Changelog.md
@@ -12,6 +12,7 @@ Release date: not released yet
 - Polyfills:
     - Added `UIntPtr.MaxValue` polyfill property
     - Added polyfills for `LeadingZeroCount(UInt32)`, `LeadingZeroCount(UInt64)`, `Log2(UInt64)`, and `Crc32C(UInt32, Byte)` methods of `BitOperations` class
+    - Added polyfill for `ToByteArray(Boolean, Boolean)` method of `BigInteger` structure
 - Fixed issues:
     - Fixed issue with `BitOperations.Log2(UInt32)` polyfill method that returned a non-zero value for zero argument on x64 processor architecture